+++ pkg/distributor/distributor.go
@@ -1923,11 +1923,20 @@ func (d *Distributor) labelValuesCardinality(ctx context.Context, labelNames []m
 		return nil, err
 	}
 
-	// When the ingest storage is enabled a partition is owned by only 1 ingester per zone,
-	// so regardless the number of zones we have it's behaving like multi-zone is always enabled.
-	isMultiZone := d.cfg.IngestStorageConfig.Enabled || (len(replicationSets) > 0 && replicationSets[0].ZoneCount() > 1)
+	zonesTotal := 0
+	if len(replicationSets) > 0 {
+		zonesTotal = replicationSets[0].ZoneCount()
+	}
+	approximateFromZonesFunc := func(countByZone map[string]uint64) uint64 {
+		return approximateFromZones(zonesTotal, d.ingestersRing.ReplicationFactor(), countByZone)
+	}
+	// When the ingest storage is enabled a partition is owned by only 1 ingester per zone.
+	// So we always approximate the resulting stats as max of what a single zone has.
+	if d.cfg.IngestStorageConfig.Enabled {
+		approximateFromZonesFunc = maxFromZones[uint64]
+	}
 
-	return cardinalityConcurrentMap.toResponse(isMultiZone, d.ingestersRing.ReplicationFactor()), nil
+	return cardinalityConcurrentMap.toResponse(approximateFromZonesFunc), nil
 }
 
 func toLabelValuesCardinalityRequest(labelNames []model.LabelName, matchers []*labels.Matcher, countMethod cardinality.CountMethod) (*ingester_client.LabelValuesCardinalityRequest, error) {
@@ -2017,7 +2026,7 @@ func (cm *labelValuesCardinalityConcurrentMap) processMessage(replicationSetIdx 
 
 // toResponse adjust builds and returns LabelValuesCardinalityResponse containing the count of series by label name
 // and value.
-func (cm *labelValuesCardinalityConcurrentMap) toResponse(isMultiZone bool, replicationFactor int) *ingester_client.LabelValuesCardinalityResponse {
+func (cm *labelValuesCardinalityConcurrentMap) toResponse(approximateFromZonesFunc func(countByZone map[string]uint64) uint64) *ingester_client.LabelValuesCardinalityResponse {
 	// we need to acquire the lock to prevent concurrent read/write to the map
 	cm.labelValuesCountersMx.Lock()
 	defer cm.labelValuesCountersMx.Unlock()
@@ -2034,7 +2043,7 @@ func (cm *labelValuesCardinalityConcurrentMap) toResponse(isMultiZone bool, repl
 			total := uint64(0)
 
 			for _, countByZone := range dataByReplicationSet {
-				total += approximateFromZones(isMultiZone, replicationFactor, countByZone)
+				total += approximateFromZonesFunc(countByZone)
 			}
 
 			countByLabelValue[labelValue] = total
@@ -2232,37 +2241,47 @@ func (r *activeSeriesResponse) result() []labels.Labels {
 	return result
 }
 
-// approximateFromZones computes a zonal value while factoring in replication.
+// approximateFromZones computes a zonal value while factoring in replication;
 // e.g. series cardinality or ingestion rate.
 //
 // If Mimir isn't deployed in a multi-zone configuration, approximateFromZones
 // divides the sum of all values by the replication factor to come up with an approximation.
-func approximateFromZones[T ~float64 | ~uint64](isMultiZone bool, replicationFactor int, seriesCountMapByZone map[string]T) T {
+func approximateFromZones[T ~float64 | ~uint64](zonesTotal, replicationFactor int, seriesCountByZone map[string]T) T {
 	// If we have more than one zone, we return the max value across zones.
 	// Values can be different across zones due to incomplete replication or
 	// other issues. Any inconsistency should always be an underestimation of
 	// the real value, so we take the max to get the best available
 	// approximation.
-	if isMultiZone {
-		var max T
-		for _, seriesCount := range seriesCountMapByZone {
-			if seriesCount > max {
-				max = seriesCount
-			}
+	if zonesTotal > 1 {
+		val := maxFromZones(seriesCountByZone)
+		if zonesTotal <= replicationFactor {
+			return val
 		}
-		return max
+		// When the number of zones is larger than the replication factor, we factor
+		// that into the approximation. We multiply the max of all zones to a ratio
+		// of number of zones to the replication factor to approximate how series
+		// are spread across zones.
+		return T(math.Round(float64(val) * float64(zonesTotal) / float64(replicationFactor)))
 	}
 
-	// If we have a single zone: we can't return the value directly because
-	// series will be replicated randomly across ingesters, and there's no way
-	// here to know how many unique series really exist. In this case, dividing
-	// by the replication factor will give us an approximation of the real
-	// value.
+	// If we have a single zone or number of zones is larger than RF, we can't return
+	// the value directly. The series will be replicated randomly across ingesters, and there's no way
+	// here to know how many unique series really exist. In this case, dividing by the replication factor
+	// will give us an approximation of the real value.
 	var sum T
-	for _, seriesCount := range seriesCountMapByZone {
+	for _, seriesCount := range seriesCountByZone {
 		sum += seriesCount
 	}
 	return T(math.Round(float64(sum) / float64(replicationFactor)))
+}
+
+func maxFromZones[T ~float64 | ~uint64](seriesCountByZone map[string]T) (val T) {
+	for _, seriesCount := range seriesCountByZone {
+		if seriesCount > val {
+			val = seriesCount
+		}
+	}
+	return val
 }
 
 // LabelNames returns the names of all labels from series with samples timestamp between from and to, and matching
@@ -2468,14 +2487,19 @@ func (d *Distributor) UserStats(ctx context.Context, countMethod cardinality.Cou
 			zoneNumSeries[r.zone] += r.resp.NumSeries
 		}
 
-		// When the ingest storage is enabled a partition is owned by only 1 ingester per zone,
-		// so regardless the number of zones we have it's behaving like multi-zone is always enabled.
-		isMultiZone := d.cfg.IngestStorageConfig.Enabled || replicationSet.ZoneCount() > 1
-
-		totalStats.IngestionRate += approximateFromZones(isMultiZone, d.ingestersRing.ReplicationFactor(), zoneIngestionRate)
-		totalStats.APIIngestionRate += approximateFromZones(isMultiZone, d.ingestersRing.ReplicationFactor(), zoneAPIIngestionRate)
-		totalStats.RuleIngestionRate += approximateFromZones(isMultiZone, d.ingestersRing.ReplicationFactor(), zoneRuleIngestionRate)
-		totalStats.NumSeries += approximateFromZones(isMultiZone, d.ingestersRing.ReplicationFactor(), zoneNumSeries)
+		// When the ingest storage is enabled, a partition is owned by only 1 ingester per zone.
+		// So we always approximate the resulting stats as max of what a single zone has.
+		if d.cfg.IngestStorageConfig.Enabled {
+			totalStats.IngestionRate += maxFromZones(zoneIngestionRate)
+			totalStats.APIIngestionRate += maxFromZones(zoneAPIIngestionRate)
+			totalStats.RuleIngestionRate += maxFromZones(zoneRuleIngestionRate)
+			totalStats.NumSeries += maxFromZones(zoneNumSeries)
+		} else {
+			totalStats.IngestionRate += approximateFromZones(replicationSet.ZoneCount(), d.ingestersRing.ReplicationFactor(), zoneIngestionRate)
+			totalStats.APIIngestionRate += approximateFromZones(replicationSet.ZoneCount(), d.ingestersRing.ReplicationFactor(), zoneAPIIngestionRate)
+			totalStats.RuleIngestionRate += approximateFromZones(replicationSet.ZoneCount(), d.ingestersRing.ReplicationFactor(), zoneRuleIngestionRate)
+			totalStats.NumSeries += approximateFromZones(replicationSet.ZoneCount(), d.ingestersRing.ReplicationFactor(), zoneNumSeries)
+		}
 	}
 
 	return totalStats, nil